Separated `no_severity` rule into two separate rules
