+++ templates/no_severity_ni.html
@@ -1,0 +1,31 @@
+{{ nag_preamble }}
+<p>
+    The following {{ plural('bug has', data, pword='bugs have') }} no Severity field set for the last {{ extra['nweeks'] }} {{ plural('week', extra['nweeks']) }}:
+</p>
+<table {{ table_attrs }}>
+    <thead>
+        <tr>
+            <th>Component</th>
+            <th>Bug</th>
+            <th>Summary</th>
+        </tr>
+    </thead>
+    <tbody>
+        {% for i, (product, comp, bugid, summary) in enumerate(data) -%}
+            <tr {% if i % 2 == 0 %}bgcolor="#E0E0E0"
+            {% endif -%}
+            >
+            <td>{{ product | e }}::{{ comp | e }}</td>
+            <td>
+                <a href="https://bugzilla.mozilla.org/show_bug.cgi?id={{ bugid }}">{{ bugid }}</a>
+            </td>
+            <td>{{ summary | e }}</td>
+        </tr>
+    {% endfor -%}
+</tbody>
+</table>
+{% if query_url_nag %}
+    <p>
+        See the query on <a href="{{ query_url_nag }}">Bugzilla</a>.
+    </p>
+{% endif -%}
